+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Shorthand for Reducer functions
 - Add **Deprecated** section in `CHANGELOG.md`
+- Add Redis pipeline support
 
 ### Changed